move manifest parsing to submodule
