+++ src/roman_datamodels/_stnode/_registry.py
@@ -12,6 +12,3 @@ if TYPE_CHECKING:
     from ._tagged import tagged_type
 
 NODE_CLASSES_BY_TAG: dict[str, tagged_type] = {}
-SCHEMA_URIS_BY_TAG: dict[str, str] = {}
-MANIFEST_TAG_REGISTRY: dict[str, list[str]] = {}
-TAG_MANIFEST_REGISTRY: dict[str, str] = {}
